doc: add download links to devices
The releases page of github is quite crowded, being able to get to the
correct image directly from the documentation should make it
significantly easier to find the correct image.

This is motivated by needing to look up the url for the download and the
url for installation instructions separately every time I recommend
someone install tow-boot.